backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
python-dotenv
requests